homescreenhandler: Handle output identification better
Rather than presume that we always have a valid QScreen, add a couple of
check against an invalid ones and add a further check for testing
against remote (remoting, streaming type of outputs) and compose the
output with the one provided by the compositor, which is under
"connector-output_name", which for remoting type of outputs accounts to
"remoting-remote-X" where X is an natural number starting from 1.

This also removes a warning about unused native variable in the
HomeScreenHandler constructor.

Bug-AGL: SPEC-4897
Signed-off-by: Marius Vlad <marius.vlad@collabora.com>
Change-Id: I1557065497c1172cf8a3e723cbac8b686ffcfead