AglShellGrpcClient: Activate windows with gRPC
In order to use HomescreenHandler we need to have a way to pass it, so
extend the callback mechanism to be able to pass other objects as well.

HomescreenHandler is the one that handles, indirectly, the activation
and that keeps track of the application status. We used in the direct
wayland event handler to we're going to re-use it as well.

Bug-AGL: SPEC-4912
Signed-off-by: Marius Vlad <marius.vlad@collabora.com>
Change-Id: I6eae6a5c4e144f3024dfcad6e5e0a54a8fec78a6